requirements.txt
opencv-python==4.11.0.86
face-recognition==1.3.0
cvzone==1.6.1
supabase==2.8.1
numpy==1.26.4
Pillow==10.4.0
dlib==19.24.6
streamlit==1.35.0
imutils==0.5.4
python-dotenv